chore: added the ability to check the interaction
