+++ packages/utilities/src/guards/PermissionGuard/types/PermissionHandlerInteraction.ts
@@ -1,0 +1,7 @@
+import type { CommandInteraction, ContextMenuInteraction } from "discord.js";
+import type { SimpleCommandMessage } from "discordx";
+
+export type PermissionHandlerInteraction =
+  | CommandInteraction
+  | SimpleCommandMessage
+  | ContextMenuInteraction;
