+++ packages/utilities/src/guards/PermissionGuard/types/index.ts
@@ -1,0 +1,1 @@
+export * from "./PermissionHandlerInteraction.js";
